Now we have an analytical expression to calculate the derivative of a rotated vector wrt the rotation axis we ought to be able to use this in place of the finite difference code in the stills prediction equn parameterisation.
However it does not work. Add a function to this script to investigate
why. Suspect equation (A.3) of Sauter et al. (2014) is not correct.
Issue is that the only changes allowed to a rotation axis are those
that preserve its unit length. In other words, d[e1]/dp should be
orthogonal to [e1]. But (A.3) does not encode that constraint.