Fix problem with Sauter et al. (2014) (A.3).
The issue can be summarised as follows.

   e1 = q0 X s0u

by definition. Therefore (A.3) looks reasonable:

   d[e1]/dp = q0 X d[s0u]/dp

However, note that

   e1 = r0 X s0u

is also a valid definition of e1. Because e1 is a unit vector it
can be defined as the cross product of any unit vector (q0, r0,
or any other) that lies in the plane of q0-s0u, with s0u. However,

   r0 X d[s0u]/dp != q0 X d[s0u]/dp

so these definitions are not equivalent when it comes to taking the
derivative!

In fact we want to use the vector that forms an orthogonal set with
e1 and s0u. This is already defined in the paper as

   c0 = s0u X e1

Now we can redefine e1 as

   e1 = c0 X s0u

and take the derivative

   d[e1]/dp = c0 X d[s0u]/dp.

This commit modifies the test function to take this derivative and
proves it is correct by comparison with finite differences.