Continue investigation into derivatives in Sauter et al. (2014) Appendix A.
As before (see commit 5397) we have

  e1 = q0 X s0u

by definition. Equation (A.22) looks sensible

  d[e1]/dp = d[q0]/dp X s0u

but it does not enforce the constraint that d[e1]/dp must be orthogonal to e1
(the derivative of a unit vector must not allow the length of the vector to
change).

We can use an alternative definition for e1:

  e1 = -q0 X q1

from which we obtain

  d[e1]/dp = -d[q0]/dp X q1

This commit adds tests for an analytical calculation of (d[r]/d[e1])(d[e1]/dp)
(a vector that is needed in the stills prediction parameterisation) versus
finite differences for each of the 9 crystal parameters.