Change panel group center computation
Before, when computing things like Z-offset (offset of a panel group along the detector normal), it used the d0 vector of the panel group.  However, for some panel groups, this can be at the corner of the virtual object instead of the center.  Now, compute the average of all the panel centers in the panel group and project that point onto the plane of the fast and slow axes of the panel group.  Use that as the center for computations such as Z-offset.