Bugfix for detector_congruence.py
For determining angles between two vectors v1 and v2, use a cross product projected along a known orthogonal vector to determine if the angle is negative or not.
For example, if the vectors are the detector normal (v1) and a panel group normal vector projected onto the radial plane (v2), the angle between them is positive or negative depending on if v2 is pointing toward or away from the beam center.