Changes for asb/duelling_profiles.py
Hardcode N cell = 25 when computing Kroon-Batenburgh 2015 eqn 7, as done in that paper. Using ML mosaic domain size from Sauter 2014 results in too big of a N cell parameter which sends the interference function close to zero.

New plot. Instead of plotting in detector space plot in mm space and use a 2d histogram in mm. This shows the fine detail in a zoomed in reflection profile. In the case of the CSPAD, the pixel size compared to the reflection size results in a loss of detail in the simulated reflection profile.