Reorganize phil parameters for asb/duelling_profiles.py
Expose interference_weighting parameters ncell and domain_size_angstroms.  If both set to None, use the domain size found during refinement of ML mosaic parameters. If not available, default to Kroon-Batenburg 2015 value of 25.  Include working default of 573 angstroms for domain_size_angstroms, which is the value we have been using for some example data. Results in ncell of ~414.6 unit cells per domain.

Allow choice between square and circular beam source and expose source_dimension as either the length of a side of a square source or the radius of a circular source.

Also show direction of beam center in mm plots.