Add parameter sigma_m_rotates_relp (default = True, maintaining original behavior)
If true, sigma_m is a conical distribution of relp vectors, I.E. a cloud of vectors
centered on the vector from the origin of reciprocal space to the reciprocal lattice
point. If false, rotate the crystal itself around random axes when applying rotational
mosaicity, using sigma_m as a standard deviation for small rotation angles around
these random axes.