Updates for asb/duelling_profile
Bugfix for spherical crystal shape: forgot to multiply vector by a random scalar, which is needed to not only pick points from the surface of the crystal, but to pick points in the volume of the crystal.

Move enough printouts under the 'show' and 'debug' flags to the program will run silently if called extenerally if desired.

Return reflection explicitly for external calling programs.