Bugfix for getting a randomly oriented vector using flex.random_double_r3_rotation_matrix_arvo_1992 and unit quaternions.
Converting from a r3 rotation matrix to an axis and angle using a unit quaternion intermediate results in a non-uniform distribution, with a bias twoards vectors whose x, y and z components are positive. Instead, get the random r3 rotation matrix and multiply by a unit vector. This results in a spherically uniform distribution of vectors.

Also allow sigma_b and sigma_m to be None to allow constraining them to zero.