Script to weight delta psi by various methods for use in refinement.
Weights are added to the reflection table using the key 'delpsical.weights'. Only method now is summed_intensity: first, a wilson analysis is performed, where the binning is computed from the average unit cell of the input experiments. The delta psi weight for a reflection is the ratio of intensity.sum.value of that reflection to the mean of intensity.sum.value in that reflection's resolution bin.