Code to produce a test reflection dataset for absorption correction.
Summary pdf attached detailing my investigation using a 1.5x1x1 cuboid,
with P4 lattice. Conclusion is that, for a 'regular' setup, the
spherical harmonic absorption correction seems to do pretty good
job when combined with the scale term and a tie to stop extreme values
at the polar regions. Nothing we didn't already know, but good to confirm
that our version of the aimless code is giving sensible results.