Added option to split the dataset into a working and free set, to check for overbiasing.
With the option use_free_set=True, the data selected for scaling
will be split into a working and free set (free set default 5% of
symmetry unique groups). A free RMSD is then calculated during
refinement, to give an indication of overbiasing, but is not
currently acted upon. Currently implemented for single and multi
dataset scaling but not targeted scaling.